requirements.txt
pytest>=7.0
pytest-asyncio
aiosqlite
SQLAlchemy
aiohttp
loguru
pydantic-settings
beautifulsoup4
ollama
psutil
matplotlib>=3.5.0
pandas>=1.3.0
numpy>=1.21.0
seaborn>=0.11.0
plotly>=5.0.0
